Refactoring - rename directory
