+++ audit-log-browser-platform-docker/src/main/docker/dev-log4j.properties
@@ -260,9 +260,9 @@ log4j.logger.org.alfresco.enterprise.repo.authorization.AuthorizationsConsistenc
 #-----------------------------------------------------------------------
 # Platform module logging
 #-----------------------------------------------------------------------
-log4j.logger.jp.ricksoft.audit_log_browser.audit.dataextractor.SiteNameDataExtractor=warn
-log4j.logger.jp.ricksoft.audit_log_browser.audit.datagenerator.LoginDataGenerator=warn
-log4j.logger.jp.ricksoft.audit_log_browser.schedule.AuditlogArchiveScheduler=warn
+log4j.logger.jp.ricksoft.auditlogbrowser.audit.dataextractor.SiteNameDataExtractor=warn
+log4j.logger.jp.ricksoft.auditlogbrowser.audit.datagenerator.LoginDataGenerator=warn
+log4j.logger.jp.ricksoft.auditlogbrowser.schedule.AuditlogArchiveScheduler=warn
