+++ audit-log-browser-platform/src/main/java/jp/ricksoft/auditlogbrowser/audit/dataextractor/ContentNameDataExtractor.java
@@ -2,7 +2,7 @@
  * Copyright 2018 Ricksoft Co., Ltd.
  * All rights reserved.
  */
-package jp.ricksoft.audit_log_browser.audit.dataextractor;
+package jp.ricksoft.auditlogbrowser.audit.dataextractor;
 
 import java.io.Serializable;
 import java.util.HashMap;
