+++ audit-log-browser-platform/src/main/java/jp/ricksoft/auditlogbrowser/audit/dataextractor/UserNameDataExtractor.java
@@ -1,20 +1,22 @@
-package jp.ricksoft.audit_log_browser.audit.dataextractor;
+/*
+ * Copyright 2018 Ricksoft Co., Ltd.
+ * All rights reserved.
+ */
+package jp.ricksoft.auditlogbrowser.audit.dataextractor;
 
 import java.io.Serializable;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.alfresco.repo.audit.extractor.AbstractDataExtractor;
 import org.alfresco.service.namespace.NamespaceService;
 import org.alfresco.service.namespace.QName;
 import org.alfresco.util.PropertyCheck;
 
-public class PropertyDataExtractor extends AbstractDataExtractor {
-
+public class UserNameDataExtractor extends AbstractDataExtractor {
+    
     private NamespaceService namespaceService;
-
+    
     public void setNamespaceService(NamespaceService namespaceService) {
         this.namespaceService = namespaceService;
     }
@@ -46,24 +48,25 @@ public class PropertyDataExtractor extends AbstractDataExtractor {
         // The data type of all arguments are passed in Serializable and italready checked using "isSupported" method.
         @SuppressWarnings("unchecked")
         Map<QName, Serializable> path = (Map<QName, Serializable>)in;
-
+        
         HashMap<String, String> propMap = new HashMap<String, String>();
         path.forEach((key, value) -> {
             propMap.put(convertPrefixedQName(key), convertSerializableToString(value));
         });
-
-        List<String> results = propMap.entrySet().stream()
-                    .map(e -> String.join(" : ", e.getKey(), e.getValue()))
-                    .collect(Collectors.toList());
-
-        return (Serializable) results;
-
+        
+        String result = propMap.entrySet().stream()
+                                .filter(e -> e.getKey().equals("cm:userName"))
+                                .findFirst()
+                                .get().getValue();
+        
+        return result;
+        
     }
-
+    
     private String convertPrefixedQName(QName key) {
         return key.getPrefixedQName(namespaceService).toPrefixString();
     }
-
+    
     private String convertSerializableToString(Serializable value) {
         if (value instanceof Object) {
             return value.toString();
