+++ audit-log-browser-platform/src/main/java/jp/ricksoft/auditlogbrowser/audit/datagenerator/SimpleStringDataGenerator.java
@@ -2,7 +2,7 @@
  * Copyright 2018 Ricksoft Co., Ltd.
  * All rights reserved.
  */
-package jp.ricksoft.audit_log_browser.audit.datagenerator;
+package jp.ricksoft.auditlogbrowser.audit.datagenerator;
 
 import java.io.Serializable;
