+++ audit-log-browser-platform/src/main/java/jp/ricksoft/auditlogbrowser/schedule/AuditlogArchiveScheduler.java
@@ -2,7 +2,7 @@
  * Copyright 2018 Ricksoft Co., Ltd.
  * All rights reserved.
  */
-package jp.ricksoft.audit_log_browser.schedule;
+package jp.ricksoft.auditlogbrowser.schedule;
 
 import java.io.File;
 import java.time.LocalDate;
@@ -17,12 +17,12 @@ import org.alfresco.service.cmr.repository.NodeRef;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import jp.ricksoft.audit_log_browser.NodeRef.RepositoryFolderManager;
-import jp.ricksoft.audit_log_browser.audit.AuditLogManager;
-import jp.ricksoft.audit_log_browser.file.CSVFileManager;
-import jp.ricksoft.audit_log_browser.file.FileManager;
-import jp.ricksoft.audit_log_browser.file.ZipFileManager;
-import jp.ricksoft.audit_log_browser.util.DateUtil;
+import jp.ricksoft.auditlogbrowser.NodeRef.RepositoryFolderManager;
+import jp.ricksoft.auditlogbrowser.audit.AuditLogManager;
+import jp.ricksoft.auditlogbrowser.file.CSVFileManager;
+import jp.ricksoft.auditlogbrowser.file.FileManager;
+import jp.ricksoft.auditlogbrowser.file.ZipFileManager;
+import jp.ricksoft.auditlogbrowser.util.DateUtil;
 
 public class AuditlogArchiveScheduler {
