+++ audit-log-browser-platform/src/main/java/jp/ricksoft/auditlogbrowser/util/DateUtil.java
@@ -1,4 +1,4 @@
-package jp.ricksoft.audit_log_browser.util;
+package jp.ricksoft.auditlogbrowser.util;
 
 import java.time.Instant;
 import java.time.LocalDate;
